Use a lock hierarchy to split down engine locks for files.
Gives a speed boost for file signing and stat calls as they can be executed in
parallel. The downside is that the engine lock isn't held during these
operations, so some state is not invariant. In practice through, most code is
only looking on read-only (input) DAG data so it's not too much of a drawback.