Support --language for bison, and different file ext. for flex/bison.
For Bison, support the --language option through Language = "...". If
Language is specified, Extension is required to specify the file extension.

For Flex, support overriding the file extension through
Extension. Support for '--c++' has not been added, since it does not
seem to work with version 2.5.35 (I get an error about --c++ and
--reentrant not being allowed together, yet I am not explicitly adding
--reentrant, and there seems to be no way to disable it).