First draft of VS2010 IDE project file generation.
Solutions, projects and filter files are generated.
A directory structure is created for the files as well.
A dummy project, 00-Tundra, is the one which triggers the Tundra build
operation. All other projects are just for navigating the source code.
There are some hardcoded paths in there (all marked with FIXME).