Fixed DAG nodes were created for undesired units.
This fixes recently introduced behavior where units in non-default passes
would build every time even though they were not in the default/cmdline
unit list.

The reason was that their DAGs were created and then hooked into the
pass dependency barrier, causing them to always be a part of the build.

This change makes it so that their DAG nodes are never created instead
which is the desired behavior. It is also slightly faster to never create
them.

Added regression test for this case.

Closes GH-103.