Silence GCC 4.6.1 warnings
In the headers on Ubuntu 11.10, there are annotations that cause a
warning when the return value of write and fread are unchecked by
application code.

There is a similar warning about passing in a path shorter than PATH_MAX
to the realpath function.

These problems are addressed by wrapping the write and fread calls in
empty if statements, and by enlarging the buffer used for the path.