Fixed PCH env data was not propagated.
PCH env data was set up too late, so it had no effect on implicitly built
object files as they are already set up before the create_dag function is
called.

The PCH env setup is now done in the customize_env function which was added for this purpose.. But I forgot to do it back then!