Added DefRule support
DefRule allows the creation of custom build steps that call out to other
programs to generate/transform source files without requiring external
lua script files. You can put the DefRule call straight in the regular
build file.

DefRules can be made config invariant. This is useful if nothing about
the configurations affect generated files and it would be a waste of
time to re-run the generation for each of the configurations.

The new sample generator-simple generates a source file (by calling out
to a python script) just once regardless of how many configurations are
built.