Added ObjGroup unit
An object group is just like a static library but other units that
depend on the object group include all of its objects in the link
command line instead of the static library. This is sometimes useful
when C++ global constructors for library-type stuff must still end up in
the main program. It can also be faster as no time is spent making the
archive file on rebuild. The main drawback of the approach is that the
link command line will be much larger as it will include a potentially
very long list of object files.