Fixed long standing bug with ancestors and command line changes
Previously, the command line (action) was used to compute the each node's
GUID. That was wrong, as it would cause old nodes to be found in the
ancestor history when the command line was changed and then changed back,
causing Tundra to not rebuild files even though the command line had
changed.

This change moves the action into the input signature and only uses the
output files (or annotation for file-less nodes) to compute the GUID. This
means the right node will be found most of the time in the ancestry, and
that node will be checked for command line changes properly as the command
line is now baked into the input signature.