Added toolset support for the GNU Assembler (GAS).
This is a simple addition to make available the GNU Assembler as
a separate toolset, which can be used in tandem with GCC or other
toolsets. It works the same as the existing YASM toolset, simply
add it to your Tools configuration (eg. Tools = { "gcc", "gas" }).

The project's tundra.lua and CMakeLists.txt files were also
updated to define the _GNU_SOURCE preprocessor definition when
compiling on Linux/UNIX platforms. This was necessary to get it to
build on platforms with older versions of glibc in which the
POSIX.1-2008 interfaces aren't made available by default (for
example, PTHREAD_MUTEX_RECURSIVE, which tundra is dependent upon).
This change allows tundra to be built without trouble on a vanilla
Debian 6.0 installation. It shouldn't cause problems with non-GNU
libc implementations such as OSX or the BSDs, and it is forward
compatible with newer versions of glibc.