Add support for installing on unixes
Needed to get a little dirty, we now pass tundra's home and executable
directories to tundra.lua directly from the boot sequence in
LuaInterface.cpp.

`GetTundraHomeDirectory` is updated to test each of the following in turn
before erroring out with a hint about the TUNDRA_HOME environment
variable if none were successful.

    1: TUNDRA_HOME environment variable
    2: scripts/tundra.lua or ../scripts/tundra.lua or ...
    3: TUNDRA_SCRIPT_HOME define from build
    4: Abort with hint to fill in TUNDRA_HOME

The Makefile has been updated with `make install` and `make uninstall`
rules. Note that the DESTDIR / PREFIX must be set in the ordinary make
step too so that the system script path can be passed in.