DefRule + bison/flex cleanup
Add ability to override command line per invocation of DefRule's setup
function. If the optional dictionary member "Command" is present, it
will override the rule's default command line.

Use the above to re-implement the bison syntax as a defrule.

The flex and bison rules previously didn't require passes to be
specified, which is wrong because they generate files that get included.
Without a pass they can run after or together with compiles of files
that include those generated files. Fixed that so you now have to
specify a pass for both (can be the same for both, of course, as they
are fully independent)