Fix precompiled headers in combination with globbing
Previously (at least on Windows), one was unable to combine the use of
precompiled headers and globbing. This commit fixes two issues in that
regard:

1. Improve the accuracy of the check for "is the precompiled header
source file included in the list of sources".

2. Fix the check in generic_cpp_setup which determines whether this is THE
precompiled header source file, and if so, applies the special parameters
to it.

Also, added an example to the "pch" example, which demonstrates combining
precompiled headers with globbing. This is more of a test than an example,
and I'm in two minds about whether it's best to leave it in or remove it.