Report node change events from [pre]action and explicit inputs
When a node is found in the build state but has a different input signature, generate an 'inputSignatureChanged' event in the structured log for that node, and figure out which components of the signature actually changed.

This first pass is only a partial solution; it covers changes to the action and preaction properties of a node, as well as changes to the list of explicit inputs, and timestamp/digest changes to the explicit inputs themselves. Other components of the hash are still unaccounted for and will be added in following commits.