Emit a structured log event at startup
When launching a build, report a structured log event containing information about the DAG being used and the targets requested. This allows frontend tools consuming the structured log to locate the corresponding JSON file for the DAG, and to understand the roots in the dependency chain that Tundra is building for.