Make JsonWriter use scratch memory
Change JsonWriter so that it uses a linked list of 1KB blocks of scratch memory for building its JSON, instead of a Buffer (which uses heap memory, copies all the data to new buffers on resize, etc). Should be much faster, and means we no longer need to worry about cleaning it up.