Add --silence-if-possible flag to suppress conclusion in no-op builds
If Tundra found no nodes that needed to be run, it currently still outputs a "build success (0.07 seconds), 0 items updated" message. This is OK for a main build, but when we are running Tundra as a prepass - e.g. for Bee bootstrapping - then it's quite annoying.

This flag suppresses the conclusion line if the build succeeded with no actions performed.