Add support for VS 2019
One thing of interest is that when VS 2019 produces a solution (.sln)
file, it gives it this version heading:

Microsoft Visual Studio Solution File, Format Version 12.00

The other IDE generation files such as msvc141.lua emit higher numbers
than 12 (eg msvc141.lua emits 14.0). VS 2019 refuses to open .sln files
with numbers higher than 14.00, but it seems unlikely that the versions
between 12 to 14 make any difference. I've made the 2019 version 12.00,
for consistency with the native values produced by the IDE, but I'm not
sure if this is correct.