Fixes for high CPU counts
Allow any number of build threads.

Use _SC_NPROCESSORS_ONLN to find online CPUs only:

Some Linux kernels are configured to allow "CPU hotswap" which means
they pretend to have some huge array (128) of CPUs that might come
online. This is weird. If we ask for the number of online CPUs that
should be a saner value.

std::thread::hardware_concurrency() has the same issue, I think it uses the total
count, which seems like a pretty bad bug in itself but outside the scope
of this project.